Vastly improved logic
Decreased the number of loops by a few and created much easier to read logic. Instead of creating a separate loop each time the program now calls a single function for looping through the program to find the next line starting with a timestamp.
Could potentially still be improved but I think this is likely as far as I'm going to get without threading.